+++ examples/basic_examples/aws_sns_registration_service.py
@@ -2,9 +2,8 @@ import os
 from typing import Dict
 
 import tomodachi
-from tomodachi import aws_sns_sqs
+from tomodachi import Options, aws_sns_sqs
 from tomodachi.envelope import JsonBase
-from tomodachi.options import Options
 
 
 class AWSSNSRegistrationService(tomodachi.Service):
