+++ examples/basic_examples/aws_sns_sqs_middleware_service.py
@@ -2,10 +2,9 @@ import os
 from typing import Any, Callable, Dict
 
 import tomodachi
-from tomodachi import aws_sns_sqs, aws_sns_sqs_publish
+from tomodachi import Options, aws_sns_sqs, aws_sns_sqs_publish
 from tomodachi.discovery import AWSSNSRegistration
 from tomodachi.envelope import JsonBase
-from tomodachi.options import Options
 
 
 async def middleware_function(
