+++ examples/pubsub_example/README.rst
@@ -56,9 +56,8 @@ Full code for ``service_a.py``
     from typing import Any
 
     import tomodachi
-    from tomodachi import aws_sns_sqs, aws_sns_sqs_publish
+    from tomodachi import aws_sns_sqs, aws_sns_sqs_publish, Options
     from tomodachi.envelope import JsonBase
-    from tomodachi.options import Options
 
 
     class ServiceA(tomodachi.Service):
